+++ store/types/store.go
@@ -7,6 +7,8 @@ import (
 	abci "github.com/tendermint/tendermint/abci/types"
 	dbm "github.com/tendermint/tm-db"
 
+	tmstrings "github.com/tendermint/tendermint/libs/strings"
+
 	snapshottypes "github.com/cosmos/cosmos-sdk/snapshots/types"
 	"github.com/cosmos/cosmos-sdk/types/kv"
 )
@@ -44,6 +46,7 @@ type Queryable interface {
 
 // StoreUpgrades defines a series of transformations to apply the multistore db upon load
 type StoreUpgrades struct {
+	Added   []string      `json:"added"`
 	Renamed []StoreRename `json:"renamed"`
 	Deleted []string      `json:"deleted"`
 }
@@ -61,6 +64,14 @@ type UpgradeInfo struct {
 type StoreRename struct {
 	OldKey string `json:"old_key"`
 	NewKey string `json:"new_key"`
+}
+
+// IsDeleted returns true if the given key should be added
+func (s *StoreUpgrades) IsAdded(key string) bool {
+	if s == nil {
+		return false
+	}
+	return tmstrings.StringInSlice(key, s.Added)
 }
 
 // IsDeleted returns true if the given key should be deleted